tests(auth): separate test classes and move utils to correct folder
