+++ tests/Serilog.Ui.Web.Tests/Authorization/AuthorizationDefaultTest.cs
@@ -1,0 +1,29 @@
+﻿using FluentAssertions;
+using System.Net.Http;
+using System.Threading.Tasks;
+using Ui.Web.Tests.Utilities;
+using Xunit;
+
+namespace Serilog.Ui.Web.Tests.Authorization;
+
+[Trait("Ui-Authorization", "Web")]
+public class AuthorizationDefaultTest : IClassFixture<WebAppFactory.Default>
+{
+    private readonly HttpClient _client;
+
+    public AuthorizationDefaultTest(WebAppFactory.Default factory)
+    {
+        _client = factory.CreateClient();
+    }
+
+    [Fact]
+    public async Task Local_Requests_Are_Allowed_By_Default()
+    {
+        // Act
+        var response = await _client.GetAsync("/serilog-ui/index.html");
+        var result = response.EnsureSuccessStatusCode;
+
+        // Assert
+        result.Should().NotThrow<HttpRequestException>();
+    }
+}
