+++ tests/Serilog.Ui.Web.Tests/Utilities/Authorization/TestRequestFilters.cs
@@ -1,0 +1,39 @@
+﻿using Microsoft.AspNetCore.Http;
+using Serilog.Ui.Web;
+using Serilog.Ui.Web.Authorization;
+using System.Threading.Tasks;
+
+namespace Ui.Web.Tests.Utilities.Authorization
+{
+    internal class ForbidLocalRequestFilter : IUiAuthorizationFilter
+    {
+        public bool Authorize(HttpContext httpContext)
+        {
+            return !httpContext.Request.IsLocal();
+        }
+    }
+
+    internal class AdmitRequestFilter : IUiAuthorizationFilter
+    {
+        public bool Authorize(HttpContext httpContext)
+        {
+            return !httpContext.Request.IsLocal();
+        }
+    }
+
+    internal class ForbidLocalRequestAsyncFilter : IUiAsyncAuthorizationFilter
+    {
+        public Task<bool> AuthorizeAsync(HttpContext httpContext)
+        {
+            return Task.FromResult(!httpContext.Request.IsLocal());
+        }
+    }
+
+    internal class AdmitRequestAsyncFilter : IUiAsyncAuthorizationFilter
+    {
+        public Task<bool> AuthorizeAsync(HttpContext httpContext)
+        {
+            return Task.FromResult(true);
+        }
+    }
+}
