+++ tests/Serilog.Ui.Web.Tests/Utilities/WebSampleProgram.cs
@@ -9,9 +9,9 @@ using Serilog.Sinks.InMemory;
 using Serilog.Ui.Web;
 using Serilog.Ui.Web.Authorization;
 using Serilog.Ui.Web.Endpoints;
-using Serilog.Ui.Web.Tests.Authorization;
 using System.IO;
 using System.Threading.Tasks;
+using Ui.Web.Tests.Utilities.Authorization;
 using Ui.Web.Tests.Utilities.InMemoryDataProvider;
 
 namespace Ui.Web.Tests.Utilities;
